Student Grades
In this project, I had to create a course class abe be able to enrol a student onto a course, and also be able to create a module witha title and module code. With this a percentage mark can be awared, and also the modules can be printed. I also had t make sure there were four mdules. All together a final grade should be put together from all the different four modules and it can all be printed for the student.